fix(browser): use valid urls in Request checks (#5630)
In Deno `@sentry/browser` is almost usable, but the check `new Request('')` fails silently because Deno throws a `TypeError: invalid URL` that goes to a catch block. This changes the Request call to include a valid URL `http://www.example.com` so that `@sentry/browser` can be imported in Deno without this issue.