Fix: Cannot read properties of undefined (getsentry/sentry-replay#160)
Enable TS strict null checks. Ran `yarn tsc` and then tried to squash or describe the types better in order to make TS happy.

Mostly I added optional chaining `?.` or allowed `null | undefined` to be passed as args to methods. This should mean the code behaves the same, but TS knows what to expect more clearly.
In a few places there is more code change, for example related to timestamps. We would expect to get `0` instead of `NaN` now.

This uncovers the bug that was also reported in slack: `Cannot read properties of undefined (reading 'protocol')`. Looks like the call to `getCurrentHub().getClient()?.getDsn()` can return `undefined` and destructuring assignment has an issue with that.

Fixes getsentry/sentry-replay#159