fix(node): Avoid catching domain errors in request handler (#5627)
According to the Node.js docs, domain errors are equivalent to an uncaught exception and therefore should not really be caught except to do some cleanup and then exit the process. Not doing so causes the process to potentially linger without being able to handle any request. Currently, the behaviour is also erroneous because `next` can end up be called twice.