fix(utils): Remove Element dom type (#5675)
This PR removes the use of the `Element` DOM type from `@sentry/utils`, introduced with https://github.com/getsentry/sentry-javascript/pull/5594. This breaks `@sentry/node` users that use strict TS configs.

We have to cast the type to `any` to make sure it works in different environments, but we supply a generic argument so that users can specify what type of `Element` will be returned from `querySelector`.