feat: Fix failure case in `sendReplay`, simplify `captureReplay` (getsentry/sentry-replay#177)
* Fixed the failure case in `sendReplay()`: there were mixed usage of
using promises and the promised result for handling control flow.
Refactored so that it is properly resolving and rejecting and not
returning boolean. Previously, in `flushUpdate`, `captureReplay()` was
always getting called regardless of the result of `sendReplay`. Now it
has been changed to only call when `sendReplay()` succeeds.
* Reduced the number of retries to 3, as previously the retries were
waiting up to 5 minutes
* Combined and reduced the number of replay events created (especially
in the initial case where we always created 2 events)