ref(nextjs): Simplify adding default integrations when initializing SDK (#5702)
While working on adding a `RequestData` integration to the nextjs SDK, I found that the functions we use to set default integrations in the SDK were a little hard to parse, primarily because of a docstring that didn't make any sense and vague and/or slightly inaccurate function and variable names. We also have been making things more complicated than necessary, in that

a) The existing functions in `utils/userIntegrations.ts` use the given default integration instance if no user instance is found. Meanwhile, the two index files do the same in the case that no user integrations are provided at all. But those two conditions aren't logically independent, so by simply providing a default (empty) value for the second case, we can rely on the first check to ensure the default instance is used. (See the now-removed `if (options.integrations)` check in both of the index files.)

b) The options we want to force onto any user instance we find have been living in a nested object which really doesn’t need to be nested. (See the `Options` - now `ForcedIntegrationOptions` - type in `utils/userIntegrations.ts`.)

c) We've essentially been manually implementing `Array.find`, which there's no reason for us to do. (See the `for`-loop in `addIntegrationToArray` - now `addOrUpdateIntegrationInArray` - in `utils/userIntegrations.ts`.)

This fixes those problems, rewrites the docstring, and renames a bunch of variables. It also

- Reorganizes the tests testing the adding of default integrations and refactors them to stop depending on the order or size of the eventual `integrations` array when testing an individual integration type.
- Fixes a small regex bug in `setNestedKey`.