chore(nextjs): Remove obsolete `dataFetchers` loader (#5713)
This removes the dataFetchers loader and its associated files and dependencies, now that it's been supplanted by the proxy loader. It also changes the build config slightly - somehow, the presence of everything this PR deletes made it so the loaders files output themselves where we want them, even without the entryFileNames setting they ought to have needed. Alas, doing the deletions broke the spell, and now they won't build to the right place without entryFileNames, so this adds it. (Even though we've now reverted to what logically should be true, I'm still baffled, because what on earth was happening before?)

Note: Until we fully settle the CJS questions, I'm leaving isESM.ts in the SDK, even though after this change it will be (for the moment) unused.