feat(types): Add `TransactionNameChange` interface (#5714)
This PR adds types for the `TransactionNameChange` interface:

```ts
export interface TransactionInfo {
    source: TransactionSource;
    changes: TransactionNameChange[];
    propagations: number;
};

/**
 * Object representing metadata about when a transaction name was changed.
 */
export interface TransactionNameChange {
  /**
   * Unix timestamp when the name was changed. Same type as the start and
   * end timestamps of a transaction and span.
   */
  timestamp: number;

  /** New source applied for transaction name change */
  source: TransactionSource;

  /** Number of propagations since start of transaction */
  propagations: number;
}
```

`TransactionNameChange` is an object that describes a point in time when a transaction name was changed. It contains a Unix timestamp (matching with span/transaction start/end timestamps), the transaction source for the changed name, and the number of propagations (when dynamic sampling context was sent to another SDK - through meta tag/http header) that happened since that transaction name change.