fix(node): Remove trailing slash in `@sentry/utils` import (#5776)
This removes a trailing slash which somehow crept into an `@sentry/utils` import when the request-data-adding functions were moved back to node. As a result of the trailing slash, Rollup doesn't recognize the import as external, which leads to the inclusion of files from the utils package in the node build, as well as an incorrect file structure (the root which is used for `build/cjs` and `build/esm` becomes `packages` rather than `packages/node/src`).