fix(nextjs): Handle `pathname` being passed in object in `instrumentServer` (#5782)
In nextjs version 12.2.6, the signature of `findPageComponents` which we wrap in `instrumentServer` changed[1], from accepting individual parameters to accepting an options object. This changes our wrapping to accommodate that, and as a result fixes transactions started by `instrumentServer` being parameterized as `[ object Object ]`. For backwards compatibility with older versions of next, individual parameters also continue to be accepted.

[1] https://github.com/vercel/next.js/commit/5f95b6b05031201b04d89cee7b6161296bbe7058#diff-2b8e3503c1bebf856610a5551d9c5996435b99f45db389d4eeac1b671518875eR242-R248