fix(build): Fix regex in multiline-comment-removal Rollup plugin (#5783)
The existing regex, when presented with code like this:

```js
/* 
 * This is
 * a multiline
 * comment 
 */

const thisIs = "real code"

/* 
 * This is
 * another multiline
 * comment 
 */

const thisIsAlso = "real code"
```

matches everything between the initial `/*` on the first line and the final `*/` on the second-to-last line, and the plugin thus removes all but the last line of the file. This fixes the regex so that only the comments are removed.