ref(nextjs): Update vercel testing script (#5738)
This is a small refactor to our `set-up-branch-for-test-app-use.sh` vercel script, which prepares the current branch to be tested on vercel by deleting packages unrelated to the nextjs SDK, thereby speeding up the vercel build.

The only consequential change is a switch from specifying which packages should get deleted to specifying which packages shouldn't. (The list of deletion targets kept falling out of date, each time we added a new package. With this change, it should in theory be evergreen.)