fix(nextjs): Reverse order of checks for instrumenting server (#5828)
As it stands right now, when we check whether or not to require `instrumentServer`, our check short-circuits if the code is running on vercel. This is a problem, because the first run of `isBuild()` causes a side effect (it sets an env var) which then enables all future runs, so without that initial run, future runs fail. This reverses the order of the check, so that `isBuild()` will always run.