feat(ember): Add ember-engine-router support
When using engines and the 'ember-engines-router-service' addon, this will allow performance monitoring to work, which was previously causing an error since the ember-engines-router-service router doesn't have a 'recognize' function. This hooks into the main router for the application instead, and only setups up route change listening once for navigation transactions.