ref(node): Make `RequestData` integration default (#5980)
This adds the new `RequestData` integration to the node SDK's default integrations, in order to be able to use it in SDKs besides just nextjs. (It therefore also removes it from the nextjs SDK's specific defaults.) In cases where we haven't yet set up our request handlers to use the integration, the event processor it adds will no-op because it won't find a request object in `sdkProcessingMetadata`.