chore(ember): Show warning when using invalid config
When defining Sentry config in `ENV['@sentry/ember'].sentry`, this goes through @embroider/macros under the hood. This relies on config being serializable, which means that e.g. you cannot use regex in the config.

To make this easier to pin down (as the error message can be a bit cryptic), this now validates the passed config (based on https://github.com/embroider-build/embroider/pull/1083) and shows a helpful warning message.

Note that this also moved this around a bit, leading to config being written only once on build, which should be a bit faster.