fix(vue): Don't overwrite custom transaction names of pageload transactions (#6060)
As brought to our attention in #6048, our pageload transaction start change (which makes the transaction start earlier) introduced in #5983 caused custom transaction names to be overwritten when `beforeNavigate` is used to update the transaction name. 

This patch fixes this problem by checking for the transaction source before (not) updating the current transaction name with the resolved route name once the router's `beforeEach` hook was called. Furthermore, it adds a test to cover this case. 

See https://github.com/getsentry/sentry-javascript/pull/6048#issuecomment-1293083543 for the motivation for creating this PR.