feat(remix): Enable `RequestData` integration for server-side requests (#6007)
This enables the new `RequestData` integration in the remix SDK in order to add request data to server-side events. The integration itself is already installed automatically when the SDK initializes, because it's one of the underlying node SDK's default integrations, but without access to the request, it currently no-ops. This enables it to function as intended by storing the request object in the scope's `sdkProcessingMetadata`.